// src/app/shared/model/promotion.model.ts
export class PromotionModel {
	promotionId: string;
	description: string;
	promotionType: string;
	promotionStartDate: number;
	promotionEndDate: number;
	category: string;
	fromRegDate: number;
	toRegDate: number;
	custOffficer: string;
	sector: string;
	commissionType: string;
	discountRate: string;
	industry: string;
	target: string;
	discountType: string;
	discountTime: string;
	proTrans: string;
	proValid: string;
	nameVn: string;
	timeUpdate: number;
}